check that meta paths contain env.json (#87)
Check that meta paths contain env.json before accepting them.
This will force uenv to fall back to the unsquashed meta data when
* a user has created a dummy meta path with a reframe.yaml file for test writing
* if scratch cleanup has removed the env.json file

We still prefer the external meta path, because this lets uenv authors test and tweak the meta data without rebuilding a uenv, as part of uenv development and testing.

Fixes #85